requirements/base.txt
imagesize>=1.4,<1.5
lunr[languages]>=0.7,<0.8
markdownify>=0.11,<0.12
Mastodon.py>=1.8.1,<1.9
orjson>=3.8,<3.10
packaging>=20,<24
rich_argparse>=0.6,<1.5
python-frontmatter>=1,<2
requests>=2.31,<3
typing-extensions>=4,<5 ; python_version < '3.11'
